move decoder out into main parser module
