+++ bgpkit-parser/examples/real-time-ris-live-websocket.rs
@@ -13,7 +13,7 @@ fn main() {
     let (mut socket, _response) = connect(Url::parse(RIS_LIVE_URL).unwrap())
         .expect("Can't connect to RIS Live websocket server");
 
-    // subscribe to messages from one collector
+    // subscribe to mrt_message from one collector
     // let msg = json!({"type": "ris_subscribe", "data": {"host": "rrc21"}}).to_string();
     let msg = json!({"type": "ris_subscribe", "data": null}).to_string();
     socket.write_message(Message::Text(msg)).unwrap();
