+++ bgpkit-parser/src/parser/bgp/attributes/attr_07_18_aggregator.rs
@@ -1,4 +1,3 @@
-use crate::encoder::MrtEncode;
 use crate::models::*;
 use crate::parser::ReadUtils;
 use crate::ParserError;
