+++ bgpkit-parser/src/parser/bgp/mod.rs
@@ -1,5 +1,5 @@
 /*!
-Provides parsing of BGP messages.
+Provides parsing of BGP mrt_message.
 */
 pub mod attributes;
 pub mod messages;
