--- bgpkit-parser/src/parser/mrt/messages/mod.rs
@@ -1,3 +1,0 @@
-pub(crate) mod bgp4mp;
-pub(crate) mod table_dump_message;
-pub(crate) mod table_dump_v2_message;
